TWEAK/FIX: vanilla religion loc to overwrite file. removed graphicalcult
Our customizable localisation can overwrite vanilla entries, so we need not include the entire file.
Fixed vanilla errors for those religions entries while being at it.
removed graphicalculturetypes folder because we don't need them (yet). Those are needed for the mod "culturally different cities" that used to be part of plus. When that mod gets merged in again, these files can be re-evaluated and re-added.